+++ Functions/MultiVariableRealValuedFunctionNoDerivatives.cpp
@@ -2,7 +2,7 @@
 
 double MultiVariableRealValuedFunctionNoDerivatives::Evaluate(std::span<const double> x) const
 {
-    if (x.size() != m_dim)
+    if (static_cast<int>(x.size()) != m_dim)
         throw MyException("FunctionNoDerivativesGiven::Evaluate: input dimension incorrect");
     return (*m_function)(x);
 }
